+++ framework/db/QueryBuilder.php
@@ -65,7 +65,22 @@ class QueryBuilder extends \yii\base\Object
         'EXISTS' => 'buildExistsCondition',
         'NOT EXISTS' => 'buildExistsCondition',
     ];
-
+    /**
+     * @var array map of chars to their replacements in LIKE conditions.
+     * By default it's configured to escape `%`, `_` and `\` with `\`.
+     * @since 2.0.12.
+     */
+    protected $likeEscapingCharacterReplacements = [
+        '%' => '\%',
+        '_' => '\_',
+        '\\' => '\\\\',
+    ];
+    /**
+     * @var string|null character used to escape special characters in LIKE conditions.
+     * By default it's assumed to be `\`.
+     * @since 2.0.12
+     */
+    protected $likeEscapeCharacter = null;
 
     /**
      * Constructor.
@@ -1348,24 +1363,16 @@ class QueryBuilder extends \yii\base\Object
      *   should be applied. Note that when using an escape mapping (or the third operand is not provided),
      *   the values will be automatically enclosed within a pair of percentage characters.
      * @param array $params the binding parameters to be populated
-     * @param array|null $escapeChars an array of chars as keys and their replacements as values.
-     * By default it prepends backslashes to `%`, `_` and `\`.
-     * This parameter is available since version 2.0.12.
      * @return string the generated SQL expression
      * @throws InvalidParamException if wrong number of operands have been given.
      */
-    public function buildLikeCondition($operator, $operands, &$params, $escapeChars = null)
+    public function buildLikeCondition($operator, $operands, &$params)
     {
         if (!isset($operands[0], $operands[1])) {
             throw new InvalidParamException("Operator '$operator' requires two operands.");
         }
 
-        $escapeChars = is_array($escapeChars) ? $escapeChars : [
-            '%' => '\%',
-            '_' => '\_',
-            '\\' => '\\\\',
-        ];
-        $escape = isset($operands[2]) ? $operands[2] : $escapeChars;
+        $escape = isset($operands[2]) ? $operands[2] : $this->likeEscapingCharacterReplacements;
         unset($operands[2]);
 
         if (!preg_match('/^(AND |OR |)(((NOT |))I?LIKE)/', $operator, $matches)) {
@@ -1400,7 +1407,11 @@ class QueryBuilder extends \yii\base\Object
                 $phName = self::PARAM_PREFIX . count($params);
                 $params[$phName] = empty($escape) ? $value : ('%' . strtr($value, $escape) . '%');
             }
-            $parts[] = "$column $operator $phName";
+            $escapeSql = '';
+            if ($this->likeEscapeCharacter !== null) {
+                $escapeSql = " ESCAPE '{$this->likeEscapeCharacter}'";
+            }
+            $parts[] = "{$column} {$operator} {$phName}{$escapeSql}";
         }
 
         return implode($andor, $parts);